\name{param.pan.plm}
\alias{param.pan.plm}
\title{Param Method for the 'pan.plm' Zelig Model}
\usage{
  \method{param}{pan.plm}(obj, num=1000, ...)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the 'pan.plm' Zelig Model
}
\note{
  This method is used by the 'pan.plm' Zelig model
}
\author{
  Nathan A. Paxton \email{napaxton@post.harvard.edu}
}
