\name{pFormulaFix}
\alias{pFormulaFix}
\title{PLM HOOK
Recast the plm object from pFormula class to Formula class}
\usage{
  pFormulaFix(obj, zcall, call, ...)
}
\arguments{
  \item{obj}{...}

  \item{zcall}{..}

  \item{call}{...}
  
  \item{...}{...}
}
\value{
  ...
}
\description{
  PLM HOOK Recast the plm object from pFormula class to
  Formula class
}

