\name{pmodel.response.Formula}
\alias{pmodel.response.Formula}
\title{Create a wrapper for the pmodel.response.pFormula function}
\usage{
  pmodel.response.Formula(obj, ...)
}
\arguments{
  \item{obj}{a ``Formula'' object}

  \item{...}{parameters to pass to pmodel.response}
}
\value{
  the exact result of a call to
  pmodel.response.pFormula(obj, ...)
}
\description{
  DESCRIPTION TEXT
}

