\docType{package}
\name{ZeligPanelmodels-package}
\alias{ZeligPanelmodels}
\alias{ZeligPanelmodels-package}
\title{ZeligPanelmodels}
\description{
  \tabular{ll}{ Package: \tab ZeligPanelmodels\cr Version:
  \tab 0.1-1\cr Date: \tab 2012-02-24\cr Depends: Zelig,
  plm License: \tab GPL version 2 or newer\cr }
}
\details{
  Edit this description
}
\author{
  Nathan A. Paxton
}
\keyword{package}

